+++ NEWS.md
@@ -1,4 +1,4 @@
-# rchallenge 1.2.0.9000
+# rchallenge 1.3.0 (23-10-2016)
 - `output_dir` argument of `publish` function now defaults to `"index.html"`. Useful for hosting the challenge on a GitHub repo with Github pages.
 - `glyphicon` is defunct. use `icon` instead of glyphicon.
 - `print_readerr` displays a table.